Update Nutzungsvertrag and SEPA template PDFs
This adds the new Obmann, and fixes a few LaTeX ??s from undefined
references in footer page numbers.

Also rename "nano" housing to "embedded", as used in the rest
of the text on the housing page, and change the link to that
file in `index.md`.

The template TeX sources have been updated in the
@FunkFeuer/funkfeuer_housing repository in commits
85776d925e4b826ea4aaee5ab89a3aba960d3db1,
1e386eb0545b904d03dd586745a1fe0ebca5ef6f, and
8c892693f89e112bb6cffc168982571fd042b30d.